added range and input for filter
