+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/index.js
@@ -5,7 +5,15 @@ import './ngbGenesTableFilter.scss';
 import component from './ngbGenesTableFilter.component';
 import controller from './ngbGenesTableFilter.controller';
 
-export default angular.module('ngbGenesTableFilter', [])
+import ngbTargetGenesFilterList from './ngbTargetGenesFilterList';
+import ngbTargetGenesFilterRange from './ngbTargetGenesFilterRange';
+import ngbTargetGenesFilterInput from './ngbTargetGenesFilterInput';
+
+export default angular.module('ngbGenesTableFilter', [
+    ngbTargetGenesFilterInput,
+    ngbTargetGenesFilterList,
+    ngbTargetGenesFilterRange,
+])
     .controller(controller.UID, controller)
     .component('ngbGenesTableFilter', component)
     .name;
