+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterInput/index.js
@@ -1,0 +1,9 @@
+import angular from 'angular';
+
+import component from './ngbTargetGenesFilterInput.component';
+import controller from './ngbTargetGenesFilterInput.controller';
+
+export default angular.module('ngbTargetGenesFilterInput', [])
+    .controller(controller.UID, controller)
+    .component('ngbTargetGenesFilterInput', component)
+    .name;
