+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterInput/ngbTargetGenesFilterInput.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbTargetGenesFilterInput.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbTargetGenesFilterInput.tpl.html')
+};
