+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterInput/ngbTargetGenesFilterInput.tpl.html
@@ -1,0 +1,8 @@
+<div flex>
+    <input class="ui-grid-filter-input ngb-filter-input"
+           style="margin-bottom: 1px"
+           type="text"
+           ng-model="$ctrl.displayText"
+           ng-blur="$ctrl.onBlur()"
+           ng-keydown="$ctrl.onKeyPress($event)">
+</div>
