+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterList/index.js
@@ -1,0 +1,11 @@
+import angular from 'angular';
+
+import './ngbTargetGenesFilterList.scss';
+
+import component from './ngbTargetGenesFilterList.component';
+import controller from './ngbTargetGenesFilterList.controller';
+
+export default angular.module('ngbTargetGenesFilterList', [])
+    .controller(controller.UID, controller)
+    .component('ngbTargetGenesFilterList', component)
+    .name;
