+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterList/ngbTargetGenesFilterList.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbTargetGenesFilterList.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbTargetGenesFilterList.tpl.html')
+};
