+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterRange/index.js
@@ -1,0 +1,9 @@
+import angular from 'angular';
+
+import component from './ngbTargetGenesFilterRange.component';
+import controller from './ngbTargetGenesFilterRange.controller';
+
+export default angular.module('ngbTargetGenesFilterRange', [])
+    .controller(controller.UID, controller)
+    .component('ngbTargetGenesFilterRange', component)
+    .name;
