+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterRange/ngbTargetGenesFilterRange.component.js
@@ -1,0 +1,9 @@
+import controller from './ngbTargetGenesFilterRange.controller';
+
+export default  {
+    bindings: {
+        column: '<'
+    },
+    controller: controller.UID,
+    template: require('./ngbTargetGenesFilterRange.tpl.html')
+};
