+++ client/client/app/components/ngbTargetPanel/ngbTargetsTab/ngbTargetsForm/ngbTargetGenesTable/ngbGenesTableFilter/ngbTargetGenesFilterRange/ngbTargetGenesFilterRange.tpl.html
@@ -1,0 +1,16 @@
+<div flex>
+    <input class="ui-grid-filter-input ngb-filter-input"
+           type="text"
+           placeholder="From"
+           style="margin-bottom: 1px"
+           ng-blur="$ctrl.apply()"
+           ng-key-bind="{enter: '$ctrl.apply()'}"
+           ng-model="$ctrl.valueFrom">
+    <input class="ui-grid-filter-input ngb-filter-input"
+           type="text"
+           placeholder="To"
+           style="margin-bottom: 1px"
+           ng-blur="$ctrl.apply()"
+           ng-key-bind="{enter: '$ctrl.apply()'}"
+           ng-model="$ctrl.valueTo">
+</div>
